pyproject.toml
[project]
name = "langgraph-react-agent"
version = "0.1.0"
description = "LangGraph ReAct Agent for n8n code review workflow testing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/react_agent"]
